+++ transports/plaintext/src/error.rs
@@ -28,17 +28,40 @@ pub enum PlainTextError {
     IoError(IoError),
 
     /// Failed to parse the handshake protobuf message.
-    InvalidPayload(Option<prost::DecodeError>),
+    InvalidPayload(DecodeError),
+
+    /// Failed to parse public key from bytes in protobuf message.
+    InvalidPublicKey(libp2p_core::identity::error::DecodingError),
+
+    /// Failed to parse the [`PeerId`](libp2p_core::PeerId) from bytes in the protobuf message.
+    InvalidPeerId(libp2p_core::multihash::Error),
 
     /// The peer id of the exchange isn't consistent with the remote public key.
-    InvalidPeerId,
+    PeerIdMismatch,
+}
+
+#[derive(Debug)]
+pub struct DecodeError(prost::DecodeError);
+
+impl fmt::Display for DecodeError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.0.fmt(f)
+    }
+}
+
+impl error::Error for DecodeError {
+    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+        self.0.source()
+    }
 }
 
 impl error::Error for PlainTextError {
     fn cause(&self) -> Option<&dyn error::Error> {
         match *self {
             PlainTextError::IoError(ref err) => Some(err),
-            PlainTextError::InvalidPayload(Some(ref err)) => Some(err),
+            PlainTextError::InvalidPayload(ref err) => Some(err),
+            PlainTextError::InvalidPublicKey(ref err) => Some(err),
+            PlainTextError::InvalidPeerId(ref err) => Some(err),
             _ => None,
         }
     }
@@ -48,13 +71,12 @@ impl fmt::Display for PlainTextError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
         match self {
             PlainTextError::IoError(e) => write!(f, "I/O error: {}", e),
-            PlainTextError::InvalidPayload(protobuf_error) => match protobuf_error {
-                Some(e) => write!(f, "Protobuf error: {}", e),
-                None => f.write_str("Failed to parse one of the handshake protobuf messages"),
-            },
-            PlainTextError::InvalidPeerId => f.write_str(
+            PlainTextError::InvalidPayload(_) => f.write_str("Failed to decode protobuf"),
+            PlainTextError::PeerIdMismatch => f.write_str(
                 "The peer id of the exchange isn't consistent with the remote public key",
             ),
+            PlainTextError::InvalidPublicKey(_) => f.write_str("Failed to decode public key"),
+            PlainTextError::InvalidPeerId(_) => f.write_str("Failed to decode PeerId"),
         }
     }
 }
@@ -67,6 +89,18 @@ impl From<IoError> for PlainTextError {
 
 impl From<prost::DecodeError> for PlainTextError {
     fn from(err: prost::DecodeError) -> PlainTextError {
-        PlainTextError::InvalidPayload(Some(err))
+        PlainTextError::InvalidPayload(DecodeError(err))
+    }
+}
+
+impl From<libp2p_core::identity::error::DecodingError> for PlainTextError {
+    fn from(err: libp2p_core::identity::error::DecodingError) -> PlainTextError {
+        PlainTextError::InvalidPublicKey(err)
+    }
+}
+
+impl From<libp2p_core::multihash::Error> for PlainTextError {
+    fn from(err: libp2p_core::multihash::Error) -> PlainTextError {
+        PlainTextError::InvalidPeerId(err)
     }
 }